Review: Use runtime.GOOS on getDeviceOSType
